Generate Wrapper task
https://gradle.org/docs/current/userguide/gradle_wrapper.html

The Gradle Wrapper (henceforth referred to as the “wrapper”) is the
preferred way of starting a Gradle build. The wrapper is a batch script
on Windows, and a shell script for other operating systems. When you
start a Gradle build via the wrapper, Gradle will be automatically
downloaded and used to run the build.

The wrapper is something you should check into version control. By
distributing the wrapper with your project, anyone can work with it
without needing to install Gradle beforehand. Even better, users of the
build are guaranteed to use the version of Gradle that the build was
designed to work with.

now, is possible run the application with:

./gradlew bootRun